requirements.txt
pymongo
loguru
gitpython
bashlex
yaml